fix and move loadAndWriteEnsemble
