+++ prody2/objects/data.py
@@ -399,3 +399,47 @@ class SetOfClassesTraj(SetOfClasses3D):
     ITEM_TYPE = ClassTraj
     REP_TYPE = AtomStruct
     REP_SET_TYPE = SetOfAtomStructs
+
+
+def loadAndWriteEnsemble(cls):
+    """Handle inputs to load ensemble into ProDy and write outputs"""
+
+    if isinstance(cls.inputEnsemble, Pointer):
+        inputEnsemble = [cls.inputEnsemble.get()]
+    else:
+        inputEnsemble = [ensemble.get() for ensemble in cls.inputEnsemble]
+
+    for i, ensemble in enumerate(inputEnsemble):
+        if isinstance(ensemble, SetOfAtomStructs):
+            ags = prody.parsePDB([tarStructure.getFileName() for tarStructure in ensemble])
+            ens = prody.buildPDBEnsemble(ags, match_func=prody.sameChainPos, seqid=0.,
+                                        overlap=0., superpose=False, degeneracy=cls.degeneracy.get())
+            # the ensemble gets built exactly as the input is setup and nothing gets rejected
+        else:
+            ens = inputEnsemble[i].loadEnsemble()
+
+        if i == 0:
+            cls.ens = ens
+        else:
+            cls.ens += ens
+
+    cls.ens.select(cls.selstr.get())
+
+    if os.path.exists(cls._getPath()):
+        avgStruct = cls.ens.getAtoms()
+        avgStruct.setCoords(cls.ens.getCoords())
+
+        cls.pdbFileName = cls._getPath('atoms.pdb')
+        prody.writePDB(cls.pdbFileName, avgStruct)
+        cls.averageStructure = AtomStruct()
+        cls.averageStructure.setFileName(cls.pdbFileName)
+
+        cls.dcdFileName = cls._getPath('ensemble.dcd')
+        prody.writeDCD(cls.dcdFileName, cls.ens)
+
+        cls.npzFileName = cls._getPath('ensemble.ens.npz')
+        prody.saveEnsemble(cls.ens, cls.npzFileName)
+        cls.npz = ProDyNpzEnsemble().create(cls._getPath())
+        for j in range(cls.ens.numConfs()):
+            frame = TrajFrame((j+1, cls.npzFileName), objLabel=cls.ens.getLabels()[j])
+            cls.npz.append(frame)
